refactor(security): move vault realpath write-guard to a shared util
assertWriteStaysInVault is a general untrusted-input write-containment guard,
not AI-specific. Move it from src/ai/tools/builtins/ to src/utils/ so the
package-import asset writer can reuse it. Pure relocation: sole importer
(vaultTools.ts) re-pointed, docs updated. No behavior change.
